requirements.txt
pillow
numpy

pillow >= 10.4.0
numpy >= 1.26.3
